requirements.txt
starlette==0.12.9
uvicorn==0.10.9
pydantic==1.2
fastapi==0.45.0
python-dotenv==0.10.3
SQLAlchemy==1.3.17
sqlalchemy-utils==0.36.6
psycopg2==2.8.5
requests==2.23.0
gunicorn